requirements.txt
pyautogui==0.9.54
pytesseract==0.3.13
python-dotenv==1.2.1
